refactor(frontend): move getSplTransactions to service modules (#4578)
# Motivation

Similar to `getSolTransactions`, we move `getSplTransactions` to the
service modules.
